Hardcode AI images sha on ImageSet
The SHA of the images that we empirically figured that AI uses
do not correspond to the ones that oc-mirror/skopeo/opm uses.
So it seems that AI has hardcoded SHAs for those images, and then
they pushed newer versions to registry sharing same labels and
versions  with no bump.
Therefore, hardcoding the SHAs to match AI in practice till we
figure out a proper and maintainable way to infer them.