Add options for AI images and imageset generation
Added options to allow the user to specify the full image reference
for the Assisted Installer images, not just the SHA. This allows the
user to override the default image location/name.

Added optional parameters to generate the imageset.yaml as a separate
step, allowing the user to add to or modify the imageset.yaml file as
needed before using that file for oc-mirror.
  --generate-imageset will generate the imageset.yaml and halt
  --skip-imageset will continue the download using the
already-generated imageset.yaml

Signed-off-by: Don Penney <dpenney@redhat.com>