Replace AI image options with more generic --ai-img
Rather than explicitly define a set of Assisted Installer images that
are supported, this update adds a --ai-img option that can be used to
specify multiple AI images. Images specified with this option are
added to ai-images.txt, whereas the --img option adds to ocp-images.txt.
The assisted-installer-reporter image, however, is added to both, as a
special case.

Signed-off-by: Don Penney <dpenney@redhat.com>