Parallelize image downloads in tool
Added parallelization of image downloads using Go worker pools with a
configurable number of workers, defaulting to 80% of available CPUs.
Download failures are requeued to try again, allowing up to 3 retries,
in case the initial surge of simultaneous image download requests cause
throttling from the registry.

The number of workers created is configurable with the --parallel
option.

The downloading.md document has been updated with the new option and
corresponding changes to output.

Signed-off-by: Don Penney <dpenney@redhat.com>