Include AMQ and BMER for 4.10 and up
Added a template helper function to allow for conditionals in the
template that check the specified release to be at least a specific
Y-stream version, such as including a block if the release is 4.12 or
higher. This function has been used in the template to update the
inclusion of the AMQ and BMER operators for 4.10 and up, where the
existing inclusion checked explicitly for 4.12 only.

Signed-off-by: Don Penney <dpenney@redhat.com>