Display stderr when operator version check fails
To validate the ACM and MCE versions specified, an oc-mirror command is
executed to ensure the version specified is listed in the catalog. In
the case where the catalog is unreachable, such as an authentication
failure due to missing certificates, the stderr of the oc-mirror command
is now displayed to provide more information to the user.

Signed-off-by: Don Penney <dpenney@redhat.com>