Add support for user-specified image filters
This update adds support for a user-specified list of image filters,
provided to the factory-precaching-cli download tool as a yaml file with
a list of regular expression patterns. The file is parsed prior to
starting the download, compiling each pattern. Prior to download, the
mapping.txt file returned by the oc-mirror command is parsed, with each
image checked against the list of patterns. If any pattern matches
either the image name or tag from the mapping.txt, the image is skipped
and added to the ignored-images.txt list file.

Signed-off-by: Don Penney <dpenney@redhat.com>