Support custom partition label
Rather than restricting the disk partition used for storage of the
prestaged images and associated files to be "data", this update adds
support for a user-specified partition label, keeping "data" as the
default. The partition function now allows the user to specify the
partition label and number, in addition to device and filesystem size,
and the image extraction scripts now have a --label option to specify
the image.

The example documented ignition config overrides have also been updated
to pass in the --label option to the extract scripts in the service-unit
config, to provide easy customization by simply replacing the "--label
data" argument with the custom partition label.

Signed-off-by: Don Penney <dpenney@redhat.com>