Eliminate duplicate extract script
As the content of the extract-ai.sh and extract-ocp.sh scripts are now
the same, this update replaces them with a single extract-images.sh as
the source code. The original names are maintained in the ignition
config files, as the name is used to differentiate which installation
stage is triggering the image extraction.

Signed-off-by: Don Penney <dpenney@redhat.com>