Reduce default parallelization to 50% from 80%
The retry mechanism appears to be causing the download tool to hit the
maximum number of threads allowed (GOMAXPROCS), causing pthread_create
failures in some cases. Until this can be resolved, dropping the default
parallelization to 50% will reduce the number of active threads used for
downloads, leaving more available for the retry mechanism.

Signed-off-by: Don Penney <dpenney@redhat.com>